Always process address [bug for wakeup!]
The address processing logic needs to check for interrupts
while it's running. It technically always needs to do this,
but we got around it for the common case before by checking
whether anything had won arbitration and jumping right to
data which would catch the interrupt.

That workaround breaks, however, if you ever have a system
where you don't scope the arbitration winner.